+++ pf2rnaseq/figures/figureHeiserFMS.py
@@ -1,5 +1,6 @@
 """
 factorization score
+factorization score
 """
 
 import numpy as np
